+++ semgrep-core/src/parsing/Parse_rule.ml
@@ -523,9 +523,10 @@ and parse_formula_old env ((key, value) : key * G.expr) : R.formula_old =
       error_at_key env key
         (spf "unexpected key %s, did you mean metavariable-regex" (fst key))
   (* These keys are handled in Python *)
-  (* TODO create an empty construct instead of abusing Patterns *)
-  | "r2c-internal-project-depends-on" -> R.Patterns (t, [])
-  | "r2c-internal-patterns-from" -> R.Patterns (t, [])
+  (* TODO really we should either remove these keys before sending
+     the rules or handle them in OCaml, this is not good *)
+  | "r2c-internal-project-depends-on" -> R.PatFilteredInPythonTodo t
+  | "r2c-internal-patterns-from" -> R.PatFilteredInPythonTodo t
   | _ -> error_at_key env key (spf "unexpected key %s" (fst key))
 
 (* let extra = parse_extra env x in