+++ semgrep-core/src/parsing/tree_sitter/Parse_tree_sitter_helpers.mli
@@ -22,7 +22,8 @@ val debug_sexp_cst_after_error : Sexplib.Sexp.t -> unit
 
 (*
    Call a tree-sitter parser and then map the CST into an AST
-   with the user-provided function. Takes care of error handling.
+   with the user-provided function. Takes care of error handling but lets
+   exceptions go through.
 *)
 val wrap_parser :
   (unit -> 'cst Tree_sitter_run.Parsing_result.t) ->