+++ CHANGELOG.md
@@ -12,6 +12,9 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - Add an experimental key for internal team use:
   `r2c-internal-project-depends-on` that allows rules to filter based on the
   presence of 3rd-party dependencies at specific version ranges.
+- FOR INTERNAL USE: if the environment variable SEMGREP_R2C_INTERNAL_EXPLICIT_SEMGREPIGNORE
+  is set, then its value will be used as the path to find semgrepignore patterns, overriding
+  any existing .semgrepignore file
 
 ### Changed
 - CLI: parse errors (reported with `--verbose`) appear once per file,