+++ CHANGELOG.md
@@ -5,11 +5,18 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 ## Unreleased
 
 ### Added
+- Add an experimental key for internal team use: `r2c-internal-project-depends-on` that
+  allows rules to filter based on the presence of 3rd-party dependencies at specific 
+  version ranges.
+- Experimental support for Dockerfile syntax.
 - Add an experimental key for internal team use:
   `r2c-internal-project-depends-on` that allows rules to filter based on the
   presence of 3rd-party dependencies at specific version ranges.
 - Support nosemgrep comments placed on the line before a match,
   causing such match to be ignored (#3521)
+- FOR INTERNAL USE: if the environment variable SEMGREP_R2C_INTERNAL_EXPLICIT_SEMGREPIGNORE
+  is set, then its value will be used as the path to find semgrepignore patterns, overriding
+  any existing .semgrepignore file
 
 ### Changed
 - CLI: parse errors (reported with `--verbose`) appear once per file, 
@@ -33,6 +40,7 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 - `--verbose` outputs a timing and file breakdown summary at the end
 - `metavariable-comparison` now handles metavariables that bind to arbitrary
   constant expressions (instead of just code variables)
+- YAML support for anchors and aliases (#3677)
 
 ### Fixed
 - Rust: inner attributes are allowed again inside functions (#4444) (#4445)