+++ CHANGELOG.md
@@ -4,6 +4,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ## Unreleased
 
+## [0.79.0](https://github.com/returntocorp/semgrep/releases/tag/v0.79.0) - 01-20-2022
+
 ### Added
 - Add an experimental key for internal team use: `r2c-internal-project-depends-on` that
   allows rules to filter based on the presence of 3rd-party dependencies at specific